comic-storyboard/backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-openai
httpx
python-dotenv
pillow
